requirements.txt
Flask
flask-cors
flask-sqlalchemy
python-dotenv
pymysql
